+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 ### Added
 - New plugin `imagick`.
+- New plugin `netlify_cms`.
 - New `expires` middleware to add the `Expires` header to responses.
 
 ### Changed